requirements.txt
absl-py==0.7.1
astor==0.7.1
Django==2.1.7
django-filter==2.1.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.3
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==3.0.3
numpy==1.16.3
opencv-python==4.1.0.25
pkg-resources==0.0.0
protobuf==3.7.1
pytz==2018.9
PyYAML==5.1
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
